Use of argparser for argument management

+++ log-sender.py
@@ -7,12 +7,13 @@
 #
 ##########################################################
 
-import boto
+#import boto
 import sys
 import os
 import time
 from boto.s3.connection import S3Connection
 from boto.s3.key import Key
+import argparse
 
 # Global variables
 INPUT_QUEUE_NAME  = 'syslog-input'
@@ -34,8 +35,15 @@ def main(argv=None):
 			secret_key = line.split()[2]
 	myfile.close()
 
+	# Get args
+	args = get_args()
+
 	# Directory to watch
-	walk_dir = sys.argv[1]
+	walk_dir = args.directory
+
+	# Compression
+	compress = args.compress
+	compress_dir = args.compress_dir
 
 	# TODO : need to put this in a infinite loop
 
@@ -50,36 +58,56 @@ def main(argv=None):
 			delta = (int(now) - int(last_modification))
 			if delta > 60 :
 				print 'send %s to bucket %s/%s' % (full_path, BUCKET_NAME, relative_path)
-				#sent_to_s3(acces_key, secret_key, BUCKET_NAME, relative_path, full_path,False)
+				sent_to_s3(acces_key, secret_key, BUCKET_NAME, relative_path, full_path, compress, compress_dir)
 				
+def get_args():
+	"""
+    Use the tools.cli methods and then add a few more arguments.
+    """
+	parser = argparse.ArgumentParser(description='Sent log file to S3')
+	parser.add_argument('--directory', type=str, help='Directory to watch', required=True)
+	parser.add_argument('--compress', type=bool, required=False, help='Activate compression', default=False)
+	parser.add_argument('--compress_dir', type=str, help='Directory to use during compression', required=False, default='/tmp')
 
-def sent_to_s3(acces_key, secret_key, bucket_name, key, filename, compress):
+	args = parser.parse_args()
 
-	'''
+	return args
+
+
+def sent_to_s3(acces_key, secret_key, bucket_name, key, filename, compress, compress_dir):
+
+	"""
 	Function to sent a file to S3 bucket
-	'''
+	"""
 
 	# TODO: put this block in try
-
 	s3_conn = S3Connection(acces_key, secret_key)
 	bucket = s3_conn.get_bucket(bucket_name)
 	bucket_key = Key(bucket)
 	bucket_key.key = str(key)
-	sent_file = open(filename, 'r')
-	bucket_key.set_contents_from_file(sent_file, replace=True, rewind=True)	
-	sent_file.close()
 
-	# Delete file
-	delete_file(filename)
+	if compress:
+		print "Compression before sending"
+
+
+	else:
+		sent_file = open(filename, 'r')
+		print "Sent to S3 desactivated"
+		#bucket_key.set_contents_from_file(sent_file, replace=True, rewind=True)	
+		sent_file.close()
+
+		# Delete file
+		delete_file(filename)
 
 def delete_file(filename):
 	
-	'''
+	"""
 	Function to delete a proceed file
-	'''
+	"""
 
 	if os.path.exists(filename):
 		print "delete file %s" % filename
+		print "Delete desactivated"
 		#os.remove(filename)
 
 if __name__ == "__main__":
